requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
ping3>=4.0.0
psutil>=5.9.0
pysnmp>=4.4.12
plotly>=5.14.0
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
